moved contents of WDTZero directory one level up and deleted WDTZero directory
