--- WDTZero/keywords.txt
@@ -1,27 +1,0 @@
-
-WDTZero	KEYWORD1
-
-clear	KEYWORD2
-setup	KEYWORD2
-attachShutdown KEYWORD2
-detachShutdown KEYWORD2
-WDTZeroCounter KEYWORD2
-
-WDT_OFF	LITERAL1
-
-WDT_HARDCYCLE62m	LITERAL1
-WDT_HARDCYCLE250m	LITERAL1
-WDT_HARDCYCLE1SS	LITERAL1
-WDT_HARDCYCLE2S	LITERAL1
-WDT_HARDCYCLE4S	LITERAL1
-WDT_HARDCYCLE8S	LITERAL1
-WDT_HARDCYCLE16S	LITERAL1
-
-WDT_SOFTCYCLE8S	LITERAL1
-WDT_SOFTCYCLE16S	LITERAL1
-WDT_SOFTCYCLE32S	LITERAL1
-WDT_SOFTCYCLE1M	LITERAL1
-WDT_SOFTCYCLE2M	LITERAL1
-WDT_SOFTCYCLE4M	LITERAL1
-WDT_SOFTCYCLE8M	LITERAL1
-WDT_SOFTCYCLE16M	LITERAL1
