+++ library.properties
@@ -1,0 +1,9 @@
+name=WDTZero
+version=1.3.0
+author=JohnV
+maintainer=Arduino <info@arduino.cc>
+sentence=Allows to use the WDT functionality. For Arduino Zero, MKRZero and MKR1000 only.
+paragraph=With this library you can use the WTD peripheral of an Arduino Zero or MKR1000 in order to program resets for software-loop timeout and hardware timeout (stalled processors).
+category=Timing
+url=http://www.arduino.cc/en/Reference/WDT
+architectures=samd
